+++ WEB-INF/src/sql/playable_class.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `playable_class` (
-	`id` 	TINYINT NOT NULL,
+	`id` 	int NOT NULL,
 	`en_US`	varchar(50) NOT NULL,
 	PRIMARY KEY (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
