+++ WEB-INF/src/sql/races.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `races` (
-	`id` 	TINYINT NOT NULL,
+	`id` 	int NOT NULL,
 	`mask` 	int NOT NULL,
 	`side`	varchar(50) NOT NULL,
 	`name`	varchar(50) NOT NULL,
